requirements.txt
flask==3.0.3
playwright==1.51.0
nest_asyncio==1.6.0
pytz==2024.1
python-dotenv==1.0.1
